backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.3.0
Flake8-pyproject==1.2.3
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
